pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bedrock_agent"]

[tool.hatch.metadata]
allow-direct-references = true

[project]
name = "aws-bedrock-browser-agent"
version = "0.1.0"
description = "AI-powered attack surface management agent using AWS Bedrock and browser automation"
readme = "README.md"
license = "MIT"
requires-python = ">=3.13"
authors = [{ name = "Matthias Baetens", email = "baetensmatthias@gmail.com" }]
dependencies = [
    "inlineagent @ git+https://github.com/awslabs/amazon-bedrock-agent-samples.git#subdirectory=src/InlineAgent",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "websockets>=12.0",
    "python-dotenv>=1.0.0",
    "playwright>=1.52.0",
    "streamlit>=1.28.0",
    "boto3>=1.34.0",
]

[project.optional-dependencies]
dev = [
    "black>=25.1.0",
    "ruff>=0.8.0",
]

[project.scripts]
bedrock-browser-agent = "bedrock_agent.cli:cli_main"

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
select = ["E", "W", "F", "I"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
